feat: add `WhichSatisfy` as filter for types
Refactor filters to be under a common directory and are separated by subject type
