--- Source/aweXpect.Reflection/Extensions/FilteredExtensions.cs
@@ -1,8 +1,0 @@
-﻿namespace aweXpect.Reflection;
-
-/// <summary>
-///     Extensions on filtered collections.
-/// </summary>
-public static partial class FilteredExtensions
-{
-}
