+++ Source/aweXpect.Reflection/Filters/AssemblyFilters.cs
@@ -1,0 +1,10 @@
+﻿using aweXpect.Reflection.Collections;
+
+namespace aweXpect.Reflection;
+
+/// <summary>
+///     Extensions methods to filter <see cref="Filtered.Assemblies" />.
+/// </summary>
+public static partial class AssemblyFilters
+{
+}
