+++ Source/aweXpect.Reflection/Filters/MethodFilters.With.cs
@@ -1,53 +1,56 @@
 ﻿using System;
+using System.Reflection;
 using System.Runtime.CompilerServices;
 using aweXpect.Reflection.Collections;
 using aweXpect.Reflection.Helpers;
 
 namespace aweXpect.Reflection;
 
-public static partial class FilteredExtensions
+public static partial class MethodFilters
 {
 	private const string DirectText = "direct ";
 
 	/// <summary>
-	///     Filter for types with attribute of type <typeparamref name="TAttribute" />.
+	///     Filter for methods with attribute of type <typeparamref name="TAttribute" />.
 	/// </summary>
 	/// <remarks>
 	///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 	///     the attribute can be inherited from a base type.
 	/// </remarks>
-	public static TypesWith With<TAttribute>(this Filtered.Types @this, bool inherit = true)
+	public static MethodsWith With<TAttribute>(this Filtered.Methods @this, bool inherit = true)
 		where TAttribute : Attribute
 	{
-		IChangeableFilter<Type> filter = Filter.Suffix<Type>(type => type.HasAttribute<TAttribute>(inherit: inherit),
+		IChangeableFilter<MethodInfo> filter = Filter.Suffix<MethodInfo>(
+			methodInfo => methodInfo.HasAttribute<TAttribute>(inherit: inherit),
 			$"with {(inherit ? "" : DirectText)}{Formatter.Format(typeof(TAttribute))} ");
-		return new TypesWith(@this.Which(filter), filter);
+		return new MethodsWith(@this.Which(filter), filter);
 	}
 
 	/// <summary>
-	///     Filter for types with attribute of type <typeparamref name="TAttribute" /> that
-	///     matches the <paramref name="predicate" />.
+	///     Filter for methods with attribute of type <typeparamref name="TAttribute" /> that
+	///     match the <paramref name="predicate" />.
 	/// </summary>
 	/// <remarks>
 	///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 	///     the attribute can be inherited from a base type.
 	/// </remarks>
-	public static TypesWith With<TAttribute>(this Filtered.Types @this,
+	public static MethodsWith With<TAttribute>(this Filtered.Methods @this,
 		Func<TAttribute, bool>? predicate,
 		bool inherit = true,
 		[CallerArgumentExpression("predicate")]
 		string doNotPopulateThisValue = "")
 		where TAttribute : Attribute
 	{
-		IChangeableFilter<Type> filter = Filter.Suffix<Type>(type => type.HasAttribute(predicate, inherit),
+		IChangeableFilter<MethodInfo> filter = Filter.Suffix<MethodInfo>(
+			methodInfo => methodInfo.HasAttribute(predicate, inherit),
 			$"with {(inherit ? "" : DirectText)}{Formatter.Format(typeof(TAttribute))} matching {doNotPopulateThisValue} ");
-		return new TypesWith(@this.Which(filter), filter);
+		return new MethodsWith(@this.Which(filter), filter);
 	}
 
 	/// <summary>
-	///     Additional filters on types with an attribute.
+	///     Additional filters on methods with an attribute.
 	/// </summary>
-	public class TypesWith(Filtered.Types inner, IChangeableFilter<Type> filter) : Filtered.Types(inner)
+	public class MethodsWith(Filtered.Methods inner, IChangeableFilter<MethodInfo> filter) : Filtered.Methods(inner)
 	{
 		/// <summary>
 		///     Allow an alternative attribute of type <typeparamref name="TAttribute" />.
@@ -56,7 +59,7 @@ public static partial class FilteredExtensions
 		///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 		///     the attribute can be inherited from a base type.
 		/// </remarks>
-		public TypesWith OrWith<TAttribute>(bool inherit = true)
+		public MethodsWith OrWith<TAttribute>(bool inherit = true)
 			where TAttribute : Attribute
 		{
 			filter.UpdateFilter((result, type) => result || type.HasAttribute<TAttribute>(inherit: inherit),
@@ -73,7 +76,7 @@ public static partial class FilteredExtensions
 		///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 		///     the attribute can be inherited from a base type.
 		/// </remarks>
-		public TypesWith OrWith<TAttribute>(
+		public MethodsWith OrWith<TAttribute>(
 			Func<TAttribute, bool>? predicate,
 			bool inherit = true,
 			[CallerArgumentExpression("predicate")]
