+++ Source/aweXpect.Reflection/Filters/MethodFilters.cs
@@ -1,0 +1,10 @@
+﻿using aweXpect.Reflection.Collections;
+
+namespace aweXpect.Reflection;
+
+/// <summary>
+///     Extensions methods to filter <see cref="Filtered.Methods" />.
+/// </summary>
+public static partial class MethodFilters
+{
+}
