+++ Source/aweXpect.Reflection/Filters/TypeFilters.With.cs
@@ -1,54 +1,53 @@
 ﻿using System;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 using aweXpect.Reflection.Collections;
 using aweXpect.Reflection.Helpers;
 
 namespace aweXpect.Reflection;
 
-public static partial class FilteredExtensions
+public static partial class TypeFilters
 {
+	private const string DirectText = "direct ";
+
 	/// <summary>
-	///     Filter for methods with attribute of type <typeparamref name="TAttribute" />.
+	///     Filter for types with attribute of type <typeparamref name="TAttribute" />.
 	/// </summary>
 	/// <remarks>
 	///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 	///     the attribute can be inherited from a base type.
 	/// </remarks>
-	public static MethodsWith With<TAttribute>(this Filtered.Methods @this, bool inherit = true)
+	public static TypesWith With<TAttribute>(this Filtered.Types @this, bool inherit = true)
 		where TAttribute : Attribute
 	{
-		IChangeableFilter<MethodInfo> filter = Filter.Suffix<MethodInfo>(
-			methodInfo => methodInfo.HasAttribute<TAttribute>(inherit: inherit),
+		IChangeableFilter<Type> filter = Filter.Suffix<Type>(type => type.HasAttribute<TAttribute>(inherit: inherit),
 			$"with {(inherit ? "" : DirectText)}{Formatter.Format(typeof(TAttribute))} ");
-		return new MethodsWith(@this.Which(filter), filter);
+		return new TypesWith(@this.Which(filter), filter);
 	}
 
 	/// <summary>
-	///     Filter for methods with attribute of type <typeparamref name="TAttribute" /> that
-	///     match the <paramref name="predicate" />.
+	///     Filter for types with attribute of type <typeparamref name="TAttribute" /> that
+	///     matches the <paramref name="predicate" />.
 	/// </summary>
 	/// <remarks>
 	///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 	///     the attribute can be inherited from a base type.
 	/// </remarks>
-	public static MethodsWith With<TAttribute>(this Filtered.Methods @this,
+	public static TypesWith With<TAttribute>(this Filtered.Types @this,
 		Func<TAttribute, bool>? predicate,
 		bool inherit = true,
 		[CallerArgumentExpression("predicate")]
 		string doNotPopulateThisValue = "")
 		where TAttribute : Attribute
 	{
-		IChangeableFilter<MethodInfo> filter = Filter.Suffix<MethodInfo>(
-			methodInfo => methodInfo.HasAttribute(predicate, inherit),
+		IChangeableFilter<Type> filter = Filter.Suffix<Type>(type => type.HasAttribute(predicate, inherit),
 			$"with {(inherit ? "" : DirectText)}{Formatter.Format(typeof(TAttribute))} matching {doNotPopulateThisValue} ");
-		return new MethodsWith(@this.Which(filter), filter);
+		return new TypesWith(@this.Which(filter), filter);
 	}
 
 	/// <summary>
-	///     Additional filters on methods with an attribute.
+	///     Additional filters on types with an attribute.
 	/// </summary>
-	public class MethodsWith(Filtered.Methods inner, IChangeableFilter<MethodInfo> filter) : Filtered.Methods(inner)
+	public class TypesWith(Filtered.Types inner, IChangeableFilter<Type> filter) : Filtered.Types(inner)
 	{
 		/// <summary>
 		///     Allow an alternative attribute of type <typeparamref name="TAttribute" />.
@@ -57,7 +56,7 @@ public static partial class FilteredExtensions
 		///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 		///     the attribute can be inherited from a base type.
 		/// </remarks>
-		public MethodsWith OrWith<TAttribute>(bool inherit = true)
+		public TypesWith OrWith<TAttribute>(bool inherit = true)
 			where TAttribute : Attribute
 		{
 			filter.UpdateFilter((result, type) => result || type.HasAttribute<TAttribute>(inherit: inherit),
@@ -74,7 +73,7 @@ public static partial class FilteredExtensions
 		///     The optional parameter <paramref name="inherit" /> (default value <see langword="true" /> specifies, if
 		///     the attribute can be inherited from a base type.
 		/// </remarks>
-		public MethodsWith OrWith<TAttribute>(
+		public TypesWith OrWith<TAttribute>(
 			Func<TAttribute, bool>? predicate,
 			bool inherit = true,
 			[CallerArgumentExpression("predicate")]
