+++ Source/aweXpect.Reflection/Filters/TypeFilters.cs
@@ -1,0 +1,10 @@
+﻿using aweXpect.Reflection.Collections;
+
+namespace aweXpect.Reflection;
+
+/// <summary>
+///     Extensions methods to filter <see cref="Filtered.Types" />.
+/// </summary>
+public static partial class TypeFilters
+{
+}
