+++ Tests/aweXpect.Reflection.Tests/Filters/AssemblyFilters.NestedTypes.Tests.cs
@@ -1,0 +1,56 @@
+﻿using aweXpect.Reflection.Collections;
+
+namespace aweXpect.Reflection.Tests.Filters;
+
+public sealed partial class AssemblyFilters
+{
+	public sealed class NestedTypes
+	{
+		public sealed class Tests
+		{
+			[Fact]
+			public async Task ShouldApplyFilterForBaseType()
+			{
+				Filtered.Types types = In.AllLoadedAssemblies().NestedTypes();
+
+				await That(types).All().Satisfy(t => t.IsNested);
+			}
+
+			[Fact]
+			public async Task ShouldIncludeNestedInformationInErrorMessage()
+			{
+				async Task Act()
+					=> await That(In.AllLoadedAssemblies().NestedTypes())
+						.AreNotNested();
+
+				await That(Act).ThrowsException()
+					.WithMessage("""
+					             Expected that nested types in all loaded assemblies
+					             are all not nested,
+					             but it contained nested types [
+					               *
+					             ]
+					             """).AsWildcard();
+			}
+
+			[Theory]
+			[MemberData(nameof(GetAccessModifiers), MemberType = typeof(AssemblyFilters))]
+			public async Task WithAccessModifier_ShouldIncludeNestedInformationInErrorMessage(
+				AccessModifiers accessModifier, string expectedString)
+			{
+				async Task Act()
+					=> await That(In.AllLoadedAssemblies().NestedTypes(accessModifier))
+						.AreNotNested();
+
+				await That(Act).ThrowsException()
+					.WithMessage($"""
+					              Expected that {expectedString}nested types in all loaded assemblies
+					              are all not nested,
+					              but it contained nested types [
+					                *
+					              ]
+					              """).AsWildcard();
+			}
+		}
+	}
+}
