+++ Tests/aweXpect.Reflection.Tests/Filters/TypeFilters.cs
@@ -1,0 +1,5 @@
+﻿namespace aweXpect.Reflection.Tests.Filters;
+
+public sealed partial class TypeFilters
+{
+}
